Here we go. kio_tar now uses KTar for testing and listing directories. This allows browsing in tar files (not tar.gz yet, for some reason), and even looking at text files inside a tar archive.
The reason to use KTar and not the tar executable, even if KTar doesn't
support all tar hacks, is that it will allow to add support for non-local
tar files later (like tar files over ftp, or tar inside tar).
For that we'll need a KTar that is KIOJob enabled...
It also fixes the problems with tar programs having a different output when listing.

svn path=/trunk/kdebase/kioslave/; revision=35887