Committed patch by Michael Haeckel to re-enable auth support for IMAP without using the now defunct ksasl library.
This does >*NOT*< mean that this added code work.. I don't know if it
does or not.  It is being committed for the sole reason that it's
necessary for this slave to compile and it doesn't affect "normal"
usage of the slave.

Somebody (several somebodies) need to test this on their existing
setups to verify that it does indeed work as expected.

svn path=/trunk/kdebase/kioslave/; revision=103332