-use the INFOPATH envvar -work with url's like info:make (i.e. without "/") -better error message, but still not translatable :-/ -defaultMimetype=text/html -listDir is no longer supported, what for ? -qoute the topic name using "\Q$Base\E" , is this correct ?
Two problems remain: when entering "info:" in konqy (without pressing enter)
konqy tries to listDir() info:, what is not supported anymore. Konqy doesn't
seem to do this e.g. for the man ioslave. Why does it do this ? Where is
this code located ? Can it be turned off ?
And the second, what magoc turns info:/ to info:/dir when entered in konqy ?
It isn't done by the ioslave.

Bye
Alex

svn path=/trunk/kdebase/kioslave/; revision=113968