support the "\(bu" tag (and probably all tags listed in standardchar too).
IMO it looks better this way:
-    { V('b','u'), 1, "*", NULL  },
+    { V('b','u'), 1, "<li></li>", NULL  },

The assignment to chardef in expand_char was more or less braindead. chardef
appears except in expand_char() only here:
    cursor = chardef;
    while (cursor) {
        chardef = cursor->next;
        delete [] cursor->st;
        delete cursor;
        cursor = chardef;
    }
    chardef = 0;

It ain't initialized before and is set back to 0 afterwards, so in
expand_char() it was *always* 0. Feel free to correct me if I'm wrong.

 static const char *expand_char(int nr)
 {
   STRDEF *h;
-  h=chardef;
+  h=standardchar;
   if (!nr) return NULL;
   while (h)
       if (h->nr==nr) {
          curpos+=h->slen;
          return h->st;
       } else
-         h=h->next;
+        h++;
   charb[0]=nr/256;

Now I don't know of any man pages on my system which aren't displayed
correctly :-)

Alex

svn path=/trunk/kdebase/kioslave/; revision=159534