merged the backend for UPnP from branches/work into the network:/ kioslave
The backend uses the lib KUPnP, which has been pulled from playground into a
private static library again. Yet it's development should be continueed so
one day it can be also used from other programs, like those doing NAT traversal via
UPnP supporting gateways (which is evil, but needed without IPv6, I was told).

KUPnP is for now a wrapper around Coherence [C], like KDNSSD wraps Avahi.
It communicates to the Coherence server via D-Bus or is just idle, if Coherence
is not available. Other backends for KUPnP should be possible if needed, as the API
does not expose the backend. A fallback implementation using the KDE Platform is
possible (and has been started) but should not really been needed.

So, now see how many UPnP enabled devices are on your network :)

[C] http://coherence-project.org/

svn path=/trunk/KDE/kdebase/runtime/; revision=1079784