Fix tab width in text thumbnails
The width of a tab character should be set to the same size as
eight space characters.

Additionally, use a fixed-width font only when formatting with
white space is used in the middle of the text lines, making files
that use multiple spaces or tabs for identation more readable.

BUG: 249156
FIXED-IN: 4.6.0

svn path=/trunk/KDE/kdebase/runtime/; revision=1174452