Replaced    http://www.ietf.org/rfc/rfc\\{@}.txt with    http://tools.ietf.org/html/rfc\\{@}
to ensure the HTML version shown instead of the text one. This provides the
added benefit of allowing a user to navigate RFCs as properly hyperlinked
document. Moreover, those that want the simple text display can simply click
on the links provides in the HTML view to view a text or even a PDF version
of the document.

BUG: 124972
FIXED-IN: 4.7