kio_sftp: Fix use-after-free leading to data corruption.
Apply patch from Peter Klotz to correct a use-after-free situation in
kio_sftp, which causes an extra buffer of 60 kbytes to be written out
when the file size was itself a multiple of 60 kbytes.

See the bug entry for a more detailed description of the cause. A
workaround for this issue made it to KDE 4.10, this is the proper fix.

Thanks to Peter for reporting, debugging, and fixing the issue and from
everyone who reviewed the patch.

BUG:312320
FIXED-IN:4.10.1