Dont convert URLs with protocol h**p: to http:
All URL protocols with four letters starting with h and ending with p where
the middle two characters are the same are automatically converted to http.

This was introduced in r312725 to "support those strange h**p://www.kde.org
style URLs that people tend to post to message boards"

This no longer seems to be very common -> remove this functionality since
it also prevents any kioslave that uses e.g. a haab:/ or hbbp:/ protocol
from working in e.g. dolphin.

REVIEW: 118015