man: generate gperf output at build time
Look for gperf, and use it to generate the hash-based lookup for the
requests table; this replaces the static generated file in the sources,
adding a build time only dependency on gperf to the man kioslave.

The new FindGperf.cmake module is a local copy of the newly proposed
module for ECM [1], which will need to stay here until the module is
included in ECM, and the build time ECM dependency is risen to the
version containing it.

The output file is now named as header, since it is used a bit like
that, and surely is better than C++ code with a .c extension.

[1] https://phabricator.kde.org/D3830

REVIEW: 129731