Fixed unrar version check
The parameter "--version" stopped working since unrar 5.2 which was
released almost 2 years ago.

With this change this now becomes the minimum version required but
every distribution that is going to relase a version with
this commit has probably been using unrar > 5.2 for quite a while
so I don't think there is any need to provide a fallback.

BUGS: 367626
BUGS: 341305

REVIEW: 129255