Fix TagLib detection and build on Windows
Summary: Due to stricter setups on Windows, TagLib was detected but the includes were not found. This patch fixes this discrepancy. It further fixes some implicit conversions which were disallowed (from QByteArray to TagLib::Filename). End result: kio-extras builds on Windows again. No idea how long it's been like this, but it's a long time.

Test Plan:
Without patch: cmake finishes and finds taglib, build fails because includes are not found
With patch: cmake finishes and finds taglib, build competes because includes are found, and implicit conversions are now explicit.

Reviewers: #frameworks, kfunk

Reviewed By: kfunk

Subscribers: vonreth

Differential Revision: https://phabricator.kde.org/D9398