fix sftp links with new uds implementation
Summary:
When listening a directory with links in sftp, the type of the entry of a symlink was inserted twice.
Now the entry data is inserted only once.
Get rid of the goto statement using local variables.

Test Plan: Tested in sftp://127.0.0.1 with broken symbolic links and sftp://127.0.0.1/usr/lib64 with a lot of symbolic links.

Reviewers: dfaure, martinkostolny, asturmlechner, lbeltrame, bruns

Reviewed By: dfaure, martinkostolny

Differential Revision: https://phabricator.kde.org/D13475