[recentdocuments:/] Filter out files that can't be browsed with a file manager
Summary:
This is a more appropriate version of {D13797}. Basically we filter out anything that can't be browsed with a file manager, such as web linkns, `appstream://` urls, etc.

This ensures that the contents of the `recentdocuments:/` ioslave is relevant to the user.

Test Plan: Opened a link from Spectacle in my web browser. Without the patch, it shows up in `recentdocuments:/`. With the patch, it does not.

Reviewers: broulik, #dolphin, #frameworks

Subscribers: elvisangelaccio, markg

Differential Revision: https://phabricator.kde.org/D14893