Fwd port: Do not use KEncodingProber - it gives false results
forward port of 3208955e66c48b07281271933c2be5e49328720f
from old kde-runtime repo

Original commit text:
Do not use KEncodingProber - it gives false results; Try dirname or UTF8

The auto-detection of the man page file content with KEncodingProber
was not successful - there are some bug reports showing it does not work
reliable - often giving EUC-JP or gb18030 as encoding, which is wrong.

I now try to find the encoding inside the man page file
(according manconv) or from the name of the directory in which the
file resides. However, on my openSuse system, neither the definition
inside nor the directory name tells me it's UTF-8, but all pages are in
UTF-8. Therefore I now use UTF-8 as default, which can be overridden
with the env-var MAN_ICONV_INPUT_CHARSET

BUG: 321074