[Thumbnailer] Use enabled plugins passed in as metadata
The thumbnailer lives in a separate application, so it will read the enabled plugins from a different configuration than the
calling application (e.g. dolphinrc).
This leads to folder thumbnails following the default settings instead of user-configured settings, causing disabled plugins
to be used and blacklisted thumbnails never being used even if enabled by the user.

BUG: 388303

Differential Revision: https://phabricator.kde.org/D15097