apply eexist workaround from listDir to stat as well
Summary:
knetattach stats() without first doing a listDir() so in the event that
the url requires authentication it won't work and error out on the EEXISTS.
listDir has a fallback hack for this which was missing from stat.

stat now has this workaround as well and correctly brings up an auth
dialog on shares that require it.

BUG: 399699

Test Plan:
- knetattach
- smb
- data for host (folder should require auth)
- save and connect
- auth dialog comes up
- enter details
- ok
- works

Reviewers: broulik

Reviewed By: broulik

Subscribers: kde-frameworks-devel, kfm-devel

Tags: #dolphin, #frameworks

Differential Revision: https://phabricator.kde.org/D16272