Fixes memory leak of ssh_session
Summary:
Fixes memory leak of ssh_session. The mSession member variable is
allocated with a call to ssh_new in sftpOpenConnection but in the
closeConnection a call to ssh_free was missing causing the ssh_session
to be leaked.

Test Plan: Build and run kio_sftp with address sanitizer

Reviewers: #frameworks, sitter, aacid

Reviewed By: aacid

Subscribers: kde-frameworks-devel, kfm-devel

Tags: #dolphin, #frameworks

Differential Revision: https://phabricator.kde.org/D19768