Remove the man KPart, and thus the last KHtml user in kio-extras
Summary:
The man KPart was originally added to open troff files directly from
Konqueror. This however doesn't work nowadays anymore as the man page
files are compressed on most systems and thus cannot be detected as
troff files.

This does not affect the man kioslave, which remains to work as before.

Reviewers: dfaure, broulik

Reviewed By: dfaure, broulik

Subscribers: ngraham, kde-frameworks-devel, kfm-devel

Tags: #dolphin, #frameworks

Differential Revision: https://phabricator.kde.org/D23809