Improved quality of JPEG thumbnails
Summary:
BUG: 411262
FIXED-IN: 20.04

Simple tweak to increase quality of JPEG thumbnails. For some reason the thumbnailer was set to scale them down at the minimum possible quality, and this led to some rather ugly results. This is, of course, no longer the case. See the attached images for a comparison. The image was created in gimp, and shows a near worst-case for the thumbnailer before the fix.

Before:
{F7882034}

After:
{F7882035}

Reviewers: broulik, #frameworks, #vdg, cfeck, ngraham

Reviewed By: #vdg, cfeck, ngraham

Subscribers: meven, volkov, cfeck, bruns, ngraham, kde-frameworks-devel, kfm-devel

Tags: #dolphin, #frameworks

Differential Revision: https://phabricator.kde.org/D26648