smb: do not map uid and gid
Summary:
libsmbc returns the uid/gid of the process, not of the file, the data
is entirely useless to us. do not list it.
currently there is no way to get the actual uid/gid of a dirent. internally
there seems to be tech available that uses the samba unix extension but
that doesn't have any public API.

in short: we can't determine ownership client-side

BUG: 212801

Test Plan: no longer shows current user as owner. in particular when that is not true!

Reviewers: ngraham

Reviewed By: ngraham

Subscribers: kde-frameworks-devel, kfm-devel

Tags: #dolphin, #frameworks

Differential Revision: https://phabricator.kde.org/D27293