smb: retain atime properly
Summary:
this was broken since forever I guess. when the .part resume tech was
added it didn't correctly retain the access time. it was trying to
get the atime of 'file', but 'file' when resuming refers to the .part which
was deleted prior to the mtime adjustment so the QFileInfo(file).lastRead
would produce an Invalid QDateTime which then results in us setting random
nonesense as access time on the file.
instead simply use dstFile. it is the path of the actual final destination
file of which the atime is **actually** the one we want to preserve here
considering we've literally just accessed that file by copying it ^^

BUG: 410624
FIXED-IN: 19.12.3

Test Plan: copy file form smb to local has sound atime

Reviewers: ngraham, meven

Reviewed By: meven

Subscribers: kde-frameworks-devel, kfm-devel

Tags: #dolphin, #frameworks

Differential Revision: https://phabricator.kde.org/D27352