smb: throw out the windows variant of the slave
Summary:
it hasn't been built in ages because the top level cmakelists actually only
adds the smb dir when samba is found (which would be never I guess?).
it is also continuously ignored when sweeping changes are made (such as
the move to json plugin metadata or the debug category change).

overall it is entirely useless as native UNC path handling seems to work
just fine these days in for example okular and kate. so, I very much
believe that bespoke smb handling is entirely moot on windows (as it
should be, really)

Test Plan: okular and kate still open remote UNC paths without it on windows

Reviewers: ngraham, #windows, vonreth

Reviewed By: ngraham, #windows, vonreth

Subscribers: kde-frameworks-devel, kfm-devel

Tags: #dolphin, #frameworks

Differential Revision: https://phabricator.kde.org/D27647