smb: split module into two targets of which one is a static lib
Summary:
this allows us to use the intermediate static target for unit testing
(which we'll want for smburl at the very least).
the effective output remains unchanged, we simply have a fake library
now before we turn it into a proper plugin .so.

kdemain as a result was moved to its own file and only gets built into the
final plugin but not the static library as it'd not have any use there
anyway

Test Plan: builds and works still

Reviewers: ngraham, apol

Reviewed By: apol

Subscribers: apol, kde-frameworks-devel, kfm-devel

Tags: #dolphin, #frameworks

Differential Revision: https://phabricator.kde.org/D27800