smb: do not double finish
Summary:
finish() must not be called twice yet previously the case branches
would call it but then there'd also be a global finished after the switch.
remove the cases' finished().

this code is, I think, actually not used anywhere so it matters little.

Test Plan: hoping for the best!

Reviewers: ngraham

Reviewed By: ngraham

Subscribers: kde-frameworks-devel, kfm-devel

Tags: #dolphin, #frameworks

Differential Revision: https://phabricator.kde.org/D28014