add readdirplus2 support for samba>=4.12
Summary:
unlike the dirent itering we do not need a separate stat call with this
new api by instead itering on what are effectively stat-like objects.

this simply runs before the regular itering which we still use to list
shares and workgroups. browse_stat_path has been split into stat and
udsentry-construction-from-stat so it can be used across both code
paths.

fun fact on the side: if there actually was a stat() api that returned
libsmb_file_info, we could use that as well and forget about the stats
structs in general :S

BUG: 402988
FIXED-IN: 20.04.0

Test Plan:
- build samba 4.12 rc
- list server
- list share
- list dir in share

- on 4.7
- all of the above

Reviewers: ngraham, asn

Reviewed By: ngraham

Subscribers: kde-frameworks-devel, kfm-devel

Tags: #dolphin, #frameworks

Differential Revision: https://phabricator.kde.org/D27520

# Conflicts:
#	smb/CMakeLists.txt
#	smb/kio_smb.h
#	smb/kio_smb_browse.cpp