smb: add some more topical debug output on opendir failure
the output seen in https://bugs.kde.org/show_bug.cgi?id=419038 isn't
nearly useful enough.

- log the actual errnum (i.e. errno) after opendir
- if the dirfd is negative log the approximate steps to recover

specifically we end up with errnum==0 in reportError which makes no sense
whatsoever and from the output it's not currently obvious which call
chain lead to that scenario. possibly should be changed to assert even,
not doing that for now though because I want this in 20.04