smb: put a bandaid on a dangling warning I left behind
many slaves seem to do this (process bytes, then say the processed size
is the stat'd size when in fact it may not be). I'll believe it's not a
huge concern because of that, but even so there should be some indication
if that expectation fails. a warning should suffice