sftp: successfully write files to proftpd servers
proftpd doesn't like zero size writes and returns with an error leading us
to assume that the write failed, when in fact it was fine eitherway because
we didn't write anything one way or the other.
so, instead just break out of the loop early. there's nothing lost by
us not doing that final EOF write since we don't need it from a protocol
POV.
the break occurs just before writing because preceding it is "spin-up" tech
opening the file and so forth. we need to have that run even when putting
a 0b file.

this successfully writes 0B, 1B, and 128M files both with partial markers
enabled and disabled to proftpd 1.3.6c

BUG: 419999
FIXED-IN: 20.04.1