smb: resolve potential dnssd names via kdnssd
as it turns out this is vastly more efficient for our purposes. we already
know that this address should only be a dnssd address, so we can ask our
dnssd daemon directly instead of taking on the entire weight of qhostinfo
and proper DNS resolution.
asking avahi directly reduces successful resolution time down from ~4 s
to ~4 ms (yes, ms!). on resolution failure the timing remains about
the same it seems