smb: refactor auth and context logic out of slave
in an effort to implement kdirnotify for smb I needed to revisit the
ancient monolithic code structure :|

authentication is now in an authenticator class the authenticator class
is plugged into a context class and the context takes care of context
initialization (and callback handling because the smbc API is a bit meh
vis a vis callback context void*).

since the authentication now doesn't live in the slavebase anymore the
interaction with `checkCachedAuthentication()` was moved into a pure
virtual interface class. slavebase has an implementation of that
interface and so will the future kdirnotify thingy.