fixup! copy FindTaglib from ECM
update taglib includes. the previous finder made the intentional
assumption that one would include taglib/foo.h rather than foo.h. this
is not true for the new finder in ECM, so update our includes
accordingly

this unbreaks the windows build (on posixy systems /usr/include could be
in the default include list and so /usr/include/taglib would always be
found)