smb: do not throw user visible warnings about ENOENT
the user won't be able to do anything about it and the scenario can
actually appear rather easily if a file disappears between us obtaining
the dir list and the stat call (e.g. with the upcoming kdirnotify tech
that happens pretty much always).

this is now also in line with the more modern readdirplus2 which also
doesn't pay any special attention to failed stats.