Avoid massive overdrawing with sequenceIndex != 0
Commit 8ed64be94a35 ("Enlarge image in folder preview when there's only
one image") made createSubThumbnail part of drawSubThumbnail. As a side
effect, when skipValidItems is > 0, the thumbnail is not only fetched
but also drawn in the top left position.

Split fetching and drawing again.

This is visible in the file dialogs (dolphin only uses sequenceIndex==0),
when hovering a directory with the mouse.