Supply canonicalPath to thumbForDirectory instead of fake QUrl
`url` is either a `thumbnail:` URL or a local URL (when called from
the m_propagationDirectories recursion), in both cases only the local
path is relevant.

Always supply a local path to the thumbForDirectory method, this also
avoids creation of a temporary QUrl.

Make the path canonical, as the path will be used will be used for
determining the thumbname hash and the Thumb::URI metadata field.
Otherwise, the directory thumbnailer will create duplicate thumbnails
for e.g. symlinked directories:

https://specifications.freedesktop.org/thumbnail-spec/thumbnail-spec-latest.html#THUMBSAVE
> You need the absolute canonical URI for the original file,
> as stated in URI RFC 2396.